(BSR)[API] chore: Move helper functions to `core.offers.repository`
