+++ api/tests/core/offers/test_repository.py
@@ -1353,3 +1353,42 @@ class UpdateStockQuantityToDnBookedQuantityTest:
         repository.update_stock_quantity_to_dn_booked_quantity(stock.id)
         # then
         assert stock.quantity == 6
+
+
+@pytest.mark.usefixtures("db_session")
+class GetPaginatedActiveOfferIdsTest:
+    def test_limit_and_page_arguments(self):
+        venue = offerers_factories.VenueFactory()
+        offer1 = factories.OfferFactory(venue=venue)
+        offer2 = factories.OfferFactory(venue=venue)
+        offer3 = factories.OfferFactory(venue=venue)
+
+        assert repository.get_paginated_active_offer_ids(limit=2, page=0) == [offer1.id, offer2.id]
+        assert repository.get_paginated_active_offer_ids(limit=2, page=1) == [offer3.id]
+
+    def test_exclude_inactive_offers(self):
+        venue = offerers_factories.VenueFactory()
+        offer1 = factories.OfferFactory(venue=venue, isActive=True)
+        _offer2 = factories.OfferFactory(venue=venue, isActive=False)
+
+        assert repository.get_paginated_active_offer_ids(limit=2, page=0) == [offer1.id]
+
+
+@pytest.mark.usefixtures("db_session")
+class GetPaginatedOfferIdsByVenueIdTest:
+    def test_limit_and_page_arguments(self):
+        venue1 = offerers_factories.VenueFactory()
+        offer1 = factories.OfferFactory(venue=venue1)
+        offer2 = factories.OfferFactory(venue=venue1)
+        offer3 = factories.OfferFactory(venue=venue1)
+        _other_venue_offer = factories.OfferFactory()
+
+        assert repository.get_paginated_offer_ids_by_venue_id(venue1.id, limit=2, page=0) == [offer1.id, offer2.id]
+        assert repository.get_paginated_offer_ids_by_venue_id(venue1.id, limit=2, page=1) == [offer3.id]
+
+    def test_include_inactive_offers(self):
+        venue = offerers_factories.VenueFactory()
+        offer1 = factories.OfferFactory(venue=venue, isActive=True)
+        offer2 = factories.OfferFactory(venue=venue, isActive=False)
+
+        assert repository.get_paginated_offer_ids_by_venue_id(venue.id, limit=2, page=0) == [offer1.id, offer2.id]
