+++ api/tests/scripts/algolia_indexing/indexing_test.py
@@ -9,7 +9,7 @@ from pcapi.scripts.algolia_indexing.indexing import batch_indexing_venues_in_alg
 
 # FIXME (dbaty, 2021-06-25): review these tests (remove mock of database queries)
 class BatchIndexingOffersInAlgoliaFromDatabaseTest:
-    @mock.patch("pcapi.repository.offer_queries.get_paginated_active_offer_ids")
+    @mock.patch("pcapi.core.offers.repository.get_paginated_active_offer_ids")
     @mock.patch("pcapi.core.search.reindex_offer_ids")
     def test_should_index_offers_once_when_offers_per_page_is_one_and_only_one_page(
         self, mock_reindex_offer_ids, mock_get_paginated_active_offer_ids
@@ -24,7 +24,7 @@ class BatchIndexingOffersInAlgoliaFromDatabaseTest:
         assert mock_get_paginated_active_offer_ids.call_count == 2
         mock_reindex_offer_ids.assert_called_once_with([1])
 
-    @mock.patch("pcapi.repository.offer_queries.get_paginated_active_offer_ids")
+    @mock.patch("pcapi.core.offers.repository.get_paginated_active_offer_ids")
     @mock.patch("pcapi.core.search.reindex_offer_ids")
     def test_should_index_offers_twice_when_offers_per_page_is_one_and_two_pages(
         self, mock_reindex_offer_ids, mock_get_paginated_active_offer_ids
@@ -42,7 +42,7 @@ class BatchIndexingOffersInAlgoliaFromDatabaseTest:
             mock.call([2]),
         ]
 
-    @mock.patch("pcapi.repository.offer_queries.get_paginated_active_offer_ids")
+    @mock.patch("pcapi.core.offers.repository.get_paginated_active_offer_ids")
     @mock.patch("pcapi.core.search.reindex_offer_ids")
     def test_should_index_offers_from_first_page_only_when_ending_page_is_provided(
         self, mock_reindex_offer_ids, mock_get_paginated_active_offer_ids
